src/requirements.txt
Django==1.11.2
MySQL-python==1.2.5
Pillow==4.1.1
certifi==2017.4.17
chardet==3.0.4
flup==1.0.2
httplib2==0.10.3
idna==2.5
oauth2==1.9.0.post1
olefile==0.44
python-dateutil==2.6.0
python-flickr==0.3.2
python-instagram==1.3.2
pytz==2017.2
requests==2.17.3
simplejson==3.10.0
six==1.10.0
urllib3==1.21.1
wsgiref==0.1.2